+++ lib/base_flow.dart
@@ -1,5 +1,5 @@
-import 'package:integration_test_mock/network/integration_test_http_overrides.dart';
-import 'package:integration_test_mock/tester/tester_interface.dart';
+import 'package:dartvm_integration_tests/network/integration_test_http_overrides.dart';
+import 'package:dartvm_integration_tests/tester/tester_interface.dart';
 
 
 abstract class BaseFlow {