+++ lib/dartvm_integration_tests.dart
@@ -1,0 +1,7 @@
+export 'package:dartvm_integration_tests/method_channels/base_method_channel_mock.dart';
+export 'package:dartvm_integration_tests/network/integration_test_http_overrides.dart';
+export 'package:dartvm_integration_tests/network/integration_test_http_client.dart';
+export 'package:dartvm_integration_tests/tester/binding_aware_pump.dart';
+export 'package:dartvm_integration_tests/tester/tester_interface.dart';
+export 'package:dartvm_integration_tests/tester/tester_factory.dart';
+export 'package:dartvm_integration_tests/tester/patrol/patrol_tester_factory.dart';