+++ lib/method_channel_mocks.dart
@@ -1,4 +1,4 @@
-import 'package:integration_test_mock/method_channels/base_method_channel_mock.dart';
+import 'package:dartvm_integration_tests/method_channels/base_method_channel_mock.dart';
 
 class MethodChannelMocks {
   final List<BaseMethodChannelMock> channelMocks;