+++ lib/tester/patrol/patrol_tester_factory.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:integration_test_mock/tester/tester_factory.dart';
-import 'package:integration_test_mock/tester/tester_interface.dart';
+import 'package:dartvm_integration_tests/tester/tester_factory.dart';
+import 'package:dartvm_integration_tests/tester/tester_interface.dart';
 import 'package:patrol_finders/patrol_finders.dart';
 
 import 'patrol_factory_impl.dart';