+++ test/network/integration_test_http_overrides_test.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart' as http;
-import 'package:integration_test_mock/network/integration_test_http_overrides.dart';
+import 'package:dartvm_integration_tests/network/integration_test_http_overrides.dart';
 
 
 void main() {